refactor: split widget attachment and maintenance handlers
